fixed SA

+++ src/IOptimizationAlgorithms.java
@@ -2,5 +2,5 @@
 public interface IOptimizationAlgorithms {
 	public void calculate();
 	public Permutation getPermutation();
-	public float getBestResult();
+	public double getBestResult();
 }
